+++ src/app/WebProviders.tsx
@@ -6,7 +6,7 @@ import { AuthProvider } from "@/shared/lib/auth";
 import { UserProvider } from "@/shared/lib/user";
 import { CounterProvider } from "@/shared/feature-counter/context/CounterProvider";
 import { Layout } from "@/layouts";
-import { LegacyModeProvider } from "@/feature-legacy";
+import { LegacyModeProvider } from "@/web/feature-legacy";
 import ThemeRegistry from "@/app/ThemeRegistry";
 
 const queryClient = new QueryClient({
