+++ src/app/[param]/page.tsx
@@ -1,4 +1,4 @@
-import { FallbackPage } from "@/feature-legacy";
+import { FallbackPage } from "@/web/feature-legacy";
 
 const Fallback = () => {
   return <FallbackPage />;
