+++ src/app/expense-report/page.tsx
@@ -1,4 +1,4 @@
-import { ExpenseReportPage } from "@/feature-expense-report";
+import { ExpenseReportPage } from "@/web/feature-expense-report";
 
 const ExpenseReport = () => {
   return <ExpenseReportPage />;
