+++ src/app/help/page.tsx
@@ -1,4 +1,4 @@
-import { HelpPage } from "@/feature-help";
+import { HelpPage } from "@/web/feature-help";
 
 const Help = () => {
   return <HelpPage />;
