+++ src/app/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginPage } from "@/feature-login";
+import { LoginPage } from "@/web/feature-login";
 
 const Login = () => {
   return <LoginPage />;
