+++ src/web/feature-activities/components/ActivitiesPage.tsx
