+++ src/web/feature-activities/components/ActivitiesPages.styles.tsx
