+++ src/web/feature-activities/components/RangeMenuButton.tsx
@@ -1,8 +1,8 @@
-import { TriggerButton } from "@/feature-activities/components/DateMenu.styles";
+import { TriggerButton } from "@/web/feature-activities/components/DateMenu.styles";
 import Icon from "@mdi/react";
 import { mdiMenuDown } from "@mdi/js";
 import { Stack, Typography } from "@/web/lib/ui";
-import { rangeMenuItem } from "@/feature-activities/components/RangeMenu";
+import { rangeMenuItem } from "@/web/feature-activities/components/RangeMenu";
 import { MouseEvent, RefObject } from "react";
 import { useTranslation } from "@/shared/lib/i18n/client";
