+++ src/web/feature-activities/components/SearchInput.styles.tsx
