+++ src/web/feature-activities/components/SearchInput.tsx
