+++ src/web/feature-activities/components/SidebarContent.tsx
