+++ src/web/feature-counter/components/CounterPicker.tsx
@@ -6,7 +6,7 @@ import { useCounter } from "@/shared/feature-counter/context/useCounter";
 import { CounterPickerData } from "./CounterPickerData";
 import { LastCounterData } from "./LastCounterData";
 import { CircularProgress } from "@mui/joy";
-import { LoginButton } from "@/feature-login/components/LoginButton";
+import { LoginButton } from "@/web/feature-login/components/LoginButton";
 import { useTranslation } from "@/shared/lib/i18n/client";
 
 type CounterPickerProps = {
