+++ src/web/feature-expense-report/components/SidebarContent.tsx
