+++ src/web/feature-legacy/components/LegacyModeWindow.tsx
