+++ src/web/feature-legacy/hooks/useAuthChanged.ts
@@ -1,4 +1,4 @@
-import { useAuth } from "../../shared/lib/auth";
+import { useAuth } from "@/shared/lib/auth";
 import {
   useHandleWindowMessage,
   WindowMessageTypesReceived,
