+++ src/web/feature-legacy/hooks/useLegacyMode.ts
