+++ src/web/feature-legacy/hooks/useLegacyPath.ts
