+++ src/web/feature-legacy/hooks/useNotFoundPaths.ts
