+++ src/web/feature-legacy/hooks/useSetCounterInAngular.ts
